1.9.9

+++ src/net/dzikoysk/funnyguilds/basic/util/RegionUtils.java
@@ -16,23 +16,17 @@ public class RegionUtils {
 	private static File regionsFolder = new File(FunnyGuilds.getInstance().getDataFolder() + File.separator + "regions" + File.separator);
 	
 	public static Region get(String name){
-		for(Region region : regions){
-			if(region.getName().equalsIgnoreCase(name)) return region;
-		}
+		for(Region region : regions) if(region.getName().equalsIgnoreCase(name)) return region;
 		return null;
 	}
 	
 	public static boolean isIn(Location loc){
-		for(Region region : regions){
-			if(region.isIn(loc)) return true;
-		}
+		for(Region region : regions) if(region.isIn(loc)) return true;
 		return false;
 	}
 	
 	public static Region getAt(Location loc){
-		for(Region region : regions){
-			if(region.isIn(loc)) return region;
-		}
+		for(Region region : regions) if(region.isIn(loc)) return region;
 		return null;
 	}
 	
@@ -41,35 +35,26 @@ public class RegionUtils {
 			double distance = center.distance(region.getCenter());
 			int i = Config.getInstance().regionSize;
 			if(Config.getInstance().enlargeItems != null) i = Config.getInstance().enlargeItems.size()*Config.getInstance().enlargeSize + i;
-			if(distance < (2 * i + Config.getInstance().regionMinDistance)){
-				return true;
-			}
+			if(distance < (2 * i + Config.getInstance().regionMinDistance)) return true;
 		}
 		return false;	
 	}
 	
 	public static void delete(Region region){
-		if(Config.getInstance().flat){
-			File file = new File(regionsFolder, region.getName()+".yml");
-			file.delete();
-		}
-		if(Config.getInstance().mysql){
-			new DatabaseRegion(region).delete();
-		}
+		if(Config.getInstance().flat) new File(regionsFolder, region.getName()+".yml").delete();
+		if(Config.getInstance().mysql) new DatabaseRegion(region).delete();
 		region.delete();
 	}
 	
 	public static List<Region> getRegions(){
-		List<Region> x = new ArrayList<Region>(regions);
-		return x;
+		return new ArrayList<Region>(regions);
 	}
 	
 	public static List<String> getNames(List<Region> lsg){
 		List<String> list = new ArrayList<>();
 		if(lsg == null) return list;
 		for(Region r : lsg){
-			if(r == null) continue;
-			if(r.getName() != null) list.add(r.getName());
+			if(r != null && r.getName() != null) list.add(r.getName());
 		}
 		return list;
 	}
